requirements.txt
selenium>=4.15.2
openai>=1.3.7
python-dotenv>=1.0.0
webdriver_manager>=4.0.1
